server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.14.0"
django-haystack = "^3.2.1"
django-cors = "^0.1"
django-debug-toolbar = "^4.2.0"
django-filter = "^23.5"
django-extensions = "^3.2.3"
django-redis = "^5.4.0"
django-channels = "^0.7.0"
loguru = "^0.7.2"
celery = "^5.3.6"
django-celery-results = "^2.5.1"
sendgrid = "^6.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
